refactor: rename files in cache module
