+++ src/modules/Cache/Adapter/Interfaces/IAdapter.ts
