+++ src/modules/Cache/Core/IChacheCore.ts
@@ -1,0 +1,3 @@
+import { IAdapter } from "../Adapter/Interfaces/IAdapter";
+
+export type IChacheCore = IAdapter;
